+++ resources/views/tracks/show.blade.php
@@ -4,40 +4,32 @@
 	 <!-- Main jumbotron for a primary marketing message or call to action -->
     <div class="jumbotron">
         <div class="container">
-            <h1>{{ $track->name }}</h1>
+			<div class="col-md-2">
+				{{ Html::image(URL::route('image.display', $track->release->image), $track->release->name, ['class' => 'img-thumbnail img-responsive']) }}
+			</div>
+			<div class="col-md-10">
+            	<h1>{{ $track->name }} <small>{{ $track->release->name }}</small></h1>
+			</div>
         </div>
     </div>
 	<div class="container">
 		<div class="col-md-6">
 			<div class="panel panel-default">
 				<div class="panel-heading collapse navbar-collapse">
-					<h2>{{ $track->number }} - {{ $track->name }} - {{ Html::linkRoute('release.show', $track->release->name, [$track->release->id]) }}</h2>
+					<h2>{{ $track->number }} - {{ $track->name }}{{-- - {{ Html::linkRoute('release.show', $track->release->name, [$track->release->id]) }} --}}
 					@can('admin')
-					<div>
-						{!! Form::open([
-							'method' => 'DELETE',
-							'route' => ['track.destroy', $track->id],
-							'class'	=> 'navbar-form navbar-right prompt-confirm',
-							'msg'	=> trans('htmusic.are_you_sure')
-						]) !!}
-							{!! Form::submit(trans('htmusic.delete'), ['class' => 'btn btn-danger']) !!}
-						{!! Form::close() !!}
-
-						{!! Form::open([
-							'method' => 'GET',
-							'route' => ['track.edit', $track->id],
-							'class'	=> 'navbar-form navbar-right'
-						]) !!}
-							{!! Form::submit(trans('htmusic.edit'), ['class' => 'btn btn-primary']) !!}
-						{!! Form::close() !!}
+					<div class="btn-group pull-right">
+						{{ Html::linkRoute('track.edit', trans('htmusic.edit'), ['link' => $track->id], ['class' => 'btn btn-default glyphicons-edit']) }}
+						{{ Html::linkRoute('track.destroy', trans('htmusic.delete'), ['link' => $track->id], ['class' => 'btn btn-default', 'data-confirm' => trans('htmusic.are_you_sure'), 'data-token' => csrf_token(),'data-method' => 'DELETE']) }}
 					</div>
 					@endcan
+					</h2>
 				</div>
 				<div class="panel-body">
 					@if(isset($track->release->name))
 					<div class="row">
 						<div class="col-md-2">{{ trans('htmusic.release') }}:</div>
-						<div class="col-md-10">{{ $track->release->name }}</div>
+						<div class="col-md-10">{{ Html::linkRoute('release.show', $track->release->name, [$track->release->id]) }}</div>
 					</div>
 					@endif
 					@if(isset($track->length))
@@ -80,17 +72,19 @@
 
 			<div class="panel panel-default">
 				<div class="panel-heading">{{ trans('htmusic.credits') }}</div>
-				<div class="panel-body">
-					{{ $track->credit->name }}
-{{--					{{ $track->credit->credit_name }}--}}
+				{{--{{ $track->credit->name }}--}}
+				<div class="list-group">
 					@forelse ($track->credit->credit_name as $row)
-						<div class="row">
-							{{ $row->join_phrase }} {{ Html::linkRoute('artist.show', $row->artist->name, [$row->artist->id]) }} ({{ $row->work->name }})
-						</div>
+						<a href="{{ route('artist.show', $row->artist->id) }}" class="list-group-item">
+							<h4>
+								@if($row->join_phrase != '&')<small>{{ $row->join_phrase }}</small>@endif
+								{{ $row->artist->name }}
+								@if($row->work->name != 'N/A')<small>{{ $row->work->name }}</small>@endif
+							</h4>
+						</a>
 					@empty
-						<div class="row">{{ trans('htmusic.no_credits_found') }}</div>
+						<div class="list-group-item">{{ trans('htmusic.no_credits_found') }}</div>
 					@endforelse
-
 				</div>
 			</div>
 
@@ -105,20 +99,21 @@
 							{{ Html::image(URL::route('image.display', $row->id), $track->name, ['class' => 'img-thumbnail img-responsive']) }}
 						</div>
 					@empty
-						<div class="row">{{ trans('htmusic.no_images_found') }}</div>
+						<div>{{ trans('htmusic.no_images_found') }}</div>
 					@endforelse
 				</div>
 			</div>
 
 			<div class="panel panel-default">
 				<div class="panel-heading">{{ trans('htmusic.links') }}</div>
-				<div class="panel-body">
+
+				<div class="list-group">
 					@forelse ($track->links as $row)
-						<div class="col-md-3">
-							{{ Html::link( $row->url, $row->caption, ['target' => '_blank', 'title' => $row->description]) }} ({{ Html::linkRoute('link.show', trans('htmusic.view'), [$row->id])}})
-						</div>
+						<a href="{{ $row->url }}" class="list-group-item" target="_blank" title="{{ $row->description }}">
+							{{ $row->caption }}<!-- {{ $row->id }} -->
+						</a>
 					@empty
-						<div class="row">{{ trans('htmusic.no_links_found') }}</div>
+						<div class="list-group-item">{{ trans('htmusic.no_links_found') }}</div>
 					@endforelse
 				</div>
 			</div>
